Extract default database type from config.
As the bareos-config script is non-portable (bash) we better extract the
default database backend from the bareos-dir.conf. To be even more sure
the script uses the right database type the call to it should just
include the 5th argument which is the database type in the config so we
are not dependent on any default type guessing.