Implementation of LZ4 and LZ4HC compression
Currently we support ZLIB based gzip compression and LZO2 compression.
LZ4 and LZ4HC (High Compression) are like LZO2 but are much faster on
decompression and are implemented as a single source file per compressor.

As we don't want to handle things like uncompressable data etc ourself
we settled for a wrapper library which mimics the zlib API and supports
the following compression algorithms:

- FASTLZ
- LZ4
- LZ4HC

The library is available at https://github.com/exalead/fastlzlib
and the few fixes needed to make it work with Bareos are currently
in a pullrequest to the original project.

For the period until the patches merge we have a clone under the bareos project
on github.

Fixes #33: Implementation of LZ4 and LZ4HC compression

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>