Don't needlessly link binaries to compression libs.
When we link the compression libaries into the shared library
we no longer need to link the binaries against them when we
use dynamic linking. When you want static linking then you do
need to specify the compress libraries needed also for the
binary as it copies the object code into your binary.