Added the new product property "auto_config"
With the new product property you can use the new silent parameter to keep configurationsfiles during a deinstallation and a reinstall.
If it is set on false, opsi doesnt delete the configuration files and doesnt overwrite the old configuration files. The new configuration files will be saved as File.new

Signed-off-by: Marco van Wieringen <marco.van.wieringen@bareos.com>