Streamline error handling.
When we fail in backup due to a version mismatch on the FD protocol
or in restore and verify for every failure we close the FD connection
if we opened that and return from the function. For backup we keep
the wait_for_job_termination() call as we really need that but as
we closed the FD connection already when we got a FD protocol version
mismatch we no longer hang waiting for something to be returned from
the FD which won't happen when the DIR detects a protocol version
mismatch.