Plugin to do backup of Microsoft Sql Server (mssql) server
Implement a native backup plugin for MSSQL based on the
Virtual Device Interface (VDI) which uses the native MSSQL
backup tooling but instead of writing it to a tape or disk
directly it will save it to a so called virtual device from
which the plugin reads the data on restore we restore to
a virtual device and the native MSSQL software recovers
the database from that.

This is based on the mssql-fd found on sourceforge from 2010
but with a lot of changes to adhere to the Bareos coding style
and to be able to compile it with mingw instead of Visual C.

We also rewrote it to use pthreads instead of the windows threading,
and the whole code is made somewhat more modular using inline
functions which handle certain logic operations.

We also implemented the ability to overwrite the plugin options
which can be passed in via the Plugin Options method from the
restore menu.

Currently we allow on restore to restore to a so called
Point In Time (PIT) using the normal MSSQL STOPBEFORMARK, STOPAT
and STOPATMARK which allow you to rollback to a certain time
or specific LSN (Log Sequence Number). This should all make
sense to MSSQL DBA's who are used to using the native backup
tooling to make disk backups.

Fixes #199: Plugin to do backup of Microsoft Sql Server (mssql) server