Fix storage daemon plugin interface.
Instead of stealing symbols from the running storage daemon run the new
record translation via the bfuncs method array. Also don't use the RES
pointer to get the setting of the compatible setting but add an extra sd
plugin variable which can be requested by the plugin to get the exact
setting of the compatible variable.